Organise github mirrorer-related code to one file
